requirements.txt

Django==4.2
djangorestframework==3.14.0
django-filter==23.2
celery==5.3.1
redis==5.0.0
gunicorn==20.1.0
pytest==7.4.0
pytest-django==4.5.2
